+++ packages/terra-clinical-detail-view/examples/test-examples/DetailView/DetailViewDivided.example.jsx
@@ -5,12 +5,15 @@ import DetailView from '../../../lib/DetailView';
 
 const item = (<p>A Detail List Item</p>);
 const item2 = (<p>A Detail List Item to prove float concept when screen readjusts smaller.</p>);
+const accessoryElement = (<p> A sample accessory container </p>);
 
 const DetailViewDivided = () => (
   <DetailView
     title="Header"
-    subtitles={['Subtitle 1', 'Subtitle 2']}
+    secondaryTitles={['secondaryTitle1', 'secondaryTitle2']}
+    subtitles={['subtitle1', 'subTitle2']}
     graph={<div style={{ border: '1px solid black', height: '20em', width: '25em', marginBottom: '0.714rem' }}>This is where any visualizations would go</div>}
+    accessory={accessoryElement}
     details={[
       (<DetailView.DetailList title="Detail List Title" key="detailInfo-1">
         <DetailView.DetailListItem item={item} />
@@ -22,6 +25,7 @@ const DetailViewDivided = () => (
     ]}
     footer="Footer Text"
     id="DetailView"
+    isSmallerTitles={false}
   />
 );