+++ packages/terra-clinical-no-data-view/tests/nightwatch/ContentNoDataView.jsx
@@ -1,5 +1,0 @@
-import React from 'react';
-
-import NoDataView from '../../lib/NoDataView';
-
-export default () => <NoDataView subtextContent={<p>content inserted here</p>} id="NoDataView" />;