+++ packages/terra-clinical-no-data-view/tests/nightwatch/TextNoDataView.jsx
@@ -1,5 +1,0 @@
-import React from 'react';
-
-import NoDataView from '../../lib/NoDataView';
-
-export default () => <NoDataView heading="test heading" subtext="test subtext" id="NoDataView" />;