+++ packages/terra-clinical-site/src/examples/detail-view/DetailViewNoDivider.jsx
@@ -9,19 +9,21 @@ const item3 = (<LabelValueView label="Collected By" textValue="Nuse Collect" />)
 const item4 = (<LabelValueView label="Priority" textValue="High" />);
 
 const DetailViewDivided = () => (
-  <DetailView
-    title="Complete Blood Count Order"
-    subtitles={['Accession', '239493849']}
-    details={[
-      (<DetailView.DetailList title="Collection Details" key="order-info-2">
-        <DetailView.DetailListItem item={item1} />
-        <DetailView.DetailListItem item={item2} />
-        <DetailView.DetailListItem item={item3} />
-        <DetailView.DetailListItem item={item4} />
-      </DetailView.DetailList>),
-    ]}
-    isDivided={false}
-  />
+  <div style={{ border: '3px dotted #d3d4d5' }}>
+    <DetailView
+      title="Complete Blood Count Order"
+      subtitles={['Accession', '239493849']}
+      details={[
+        (<DetailView.DetailList title="Collection Details" key="order-info-2">
+          <DetailView.DetailListItem item={item1} />
+          <DetailView.DetailListItem item={item2} />
+          <DetailView.DetailListItem item={item3} />
+          <DetailView.DetailListItem item={item4} />
+        </DetailView.DetailList>),
+      ]}
+      isDivided={false}
+    />
+  </div>
 );
 
 export default DetailViewDivided;