+++ packages/terra-clinical-item-view/CHANGELOG.md
@@ -4,6 +4,16 @@ ChangeLog
 Unreleased
 ----------
 
+2.4.0 - (May 1, 2018)
+----------
+### Changed
+* Minor dependency updates
+
+2.3.0 - (April 20, 2018)
+----------
+### Changed
+* Minor dependency updates
+
 2.2.0 - (March 27, 2018)
 ----------
 ### Changed